Added Util Class
Util class will manage functions that are shared between scripts.

Added RollChance, which should be used for all luck based rolls.

Fixed shotgun killing enemies multiple times

Combined the shooting methods into one to reserve code space and simplify it

Fixed Flash Coroutine trying to activate when the object is null

+++ Assets/Scripts/Momentos/Momento.cs
@@ -15,7 +15,7 @@ public abstract class Momento : MonoBehaviour
     public virtual float GetSpreadMultiplier() => 1f;
     public virtual float GetMoveSpeedMultiplier() => 1f;
     public virtual float GetTreasureClassMultiplier() => 1f;
-    public virtual float GetLuckMultiplier() => 1f;
+    public virtual int GetLuckMultiplier() => 1;
     public virtual float GetCriticalChanceMultiplier() => 1f;
 
     /// <summary>
